+++ src/components/ContentSwitcher/ContentSwitcherComponent.ts
@@ -1,8 +1,9 @@
 import type { HTMLAttributes } from 'react';
-import styled from 'styled-components';
+import styled, { css } from 'styled-components';
 import { ContentSwitcherItemButton } from '#src/components/ContentSwitcher/ContentSwitcherItem';
 
 export type Dimension = 'l' | 'm' | 's';
+export type Adaptive = 'fill';
 
 export const SIZE = {
   l: 40,
@@ -12,6 +13,7 @@ export const SIZE = {
 
 type ContentSwitcherComponentProps = HTMLAttributes<HTMLDivElement> & {
   $dimension: Dimension;
+  $adaptive?: Adaptive;
 };
 
 const BORDER_RADIUS = {
@@ -26,6 +28,12 @@ const padding = {
   s: 12,
 };
 
+const adaptiveMixin = css`
+  width: 100%;
+  overflow: hidden;
+  min-width: fit-content;
+`;
+
 export const ContentSwitcherComponent = styled.div<ContentSwitcherComponentProps>`
   background: var(--admiral-color-Neutral_Neutral10, ${(p) => p.theme.color['Neutral/Neutral 10']});
   border: 2px solid var(--admiral-color-Neutral_Neutral10, ${(p) => p.theme.color['Neutral/Neutral 10']});
@@ -33,6 +41,8 @@ export const ContentSwitcherComponent = styled.div<ContentSwitcherComponentProps
   display: inline-flex;
   height: ${({ $dimension }) => SIZE[$dimension]}px;
   padding: 2px;
+
+  ${({ $adaptive }) => ($adaptive === 'fill' ? adaptiveMixin : '')}
 
   &[data-appearance~='primary'] {
     border: 2px solid var(--admiral-color-Primary_Primary60Main, ${(p) => p.theme.color['Primary/Primary 60 Main']});